requirements.txt
flask
pyautogui
subprocess
shlex
pydub
threading
customtkinter
Pillow
